refactor to useParams for article detail
