+++ src/Components/App/App.js
@@ -8,7 +8,6 @@ import ArticleDetail from '../ArticleDetail/ArticleDetail';
 
 function App() {
   const [articles, setArticles] = useState([])
-  const [singleArticle, setSingleArticle] = useState({})
   function getArticles() {
     const articlesWithIds = allArticles.articles.map((article, index) => {
       return {
@@ -17,12 +16,6 @@ function App() {
       }
     })
     setArticles(articlesWithIds)
-  }
-
-  function getSingleArticle(id) {
-    const article = articles.find(article => article.id === id)
-    console.log('art', article)
-    setSingleArticle(article)
   }
 
   useEffect(() => {
@@ -35,8 +28,8 @@ function App() {
     <div className="App">
       <Header />
       <Routes>
-        <Route path="/" element={<MainPage articles={articles} getSingleArticle={getSingleArticle}/>} />
-        <Route path="/:id" element={<ArticleDetail singleArticle={singleArticle} />} />
+        <Route path="/" element={<MainPage articles={articles} />} />
+        <Route path="/:id" element={<ArticleDetail articles={articles} />} />
       </Routes>
     </div>
   );
