+++ src/Components/ArticleDetail/ArticleDetail.css
@@ -1,0 +1,3 @@
+.article-detail {
+    text-align: center;
+}
